(wip) make core submodule and drop unused code
