+++ src/kimera_vio_evaluation/evaluation/metric_semantic.py
@@ -1,6 +1,7 @@
 """Run metric semantic evaluation."""
 import numpy as np
 from os import path
+import os
 import glog as log
 import copy
 
@@ -8,8 +9,6 @@ from tqdm.notebook import tqdm
 
 import open3d as o3d
 import pandas as pd
-
-from evaluation.tools.mesh import Mesh
 
 # Rotation matrices:
 # East North Up (ENU) frame to Unity's world frame of reference
@@ -19,6 +18,65 @@ unity_R_enu = np.transpose(enu_R_unity)
 # Right Handed frame to Unity's Left Handed frame of reference
 righthand_R_lefthand = np.array([[1, 0, 0], [0, -1, 0], [0, 0, 1]])
 lefthand_R_righthand = np.transpose(righthand_R_lefthand)
+
+
+class Mesh:
+    """Wrapper around an open3d mesh."""
+
+    def __init__(self, filepath):
+        """Read a mesh from file."""
+        os.path.isfile(filepath)
+        self.mesh_o3d = o3d.io.read_triangle_mesh(filepath)
+
+    def visualize(self):
+        """Draw a mesh."""
+        self.mesh_o3d.compute_vertex_normals()
+        vis = o3d.visualization.Visualizer()
+        vis.create_window()
+        vis.get_render_option().mesh_show_back_face = True
+        self.add_to_vis(vis)
+        mesh_frame = o3d.geometry.create_mesh_coordinate_frame(size=4, origin=[0, 0, 0])
+        vis.add_geometry(mesh_frame)
+        vis.run()
+        vis.destroy_window()
+
+    def add_to_vis(self, vis):
+        """Add the mesh to the visualization `vis` object."""
+        self.mesh_o3d.compute_vertex_normals()
+        vis.add_geometry(self.mesh_o3d)
+
+    def print_mesh(self):
+        """Show mesh in console."""
+        print("Testing mesh in open3d ...")
+        print(self.mesh_o3d)
+        print(np.asarray(self.mesh_o3d.vertices))
+        print(np.asarray(self.mesh_o3d.triangles))
+        print("")
+
+    def transform_left(self, rotation_matrix):
+        """Left multiply matrix."""
+        # TODO(Toni): there is a transform method!!!
+        assert isinstance(rotation_matrix, np.ndarray)
+        assert np.size(rotation_matrix, 0) == 3
+        assert np.size(rotation_matrix, 1) == 3
+        # print("Transforming mesh according to left matrix:")
+        # print(rotation_matrix)
+        rotated_vertices = rotation_matrix.dot(
+            np.transpose(np.asarray(self.mesh_o3d.vertices))
+        )
+        self.mesh_o3d.vertices = o3d.utility.Vector3dVector(
+            np.transpose(rotated_vertices)
+        )
+
+    def transform_right(self, rotation_matrix):
+        """Right multiply matrix."""
+        assert isinstance(rotation_matrix, np.ndarray)
+        assert np.size(rotation_matrix, 0) == 3
+        assert np.size(rotation_matrix, 1) == 3
+        print("Transforming mesh according to right matrix:")
+        print(rotation_matrix)
+        rotated_vertices = np.asarray(self.mesh_o3d.vertices).dot(rotation_matrix)
+        self.mesh_o3d.vertices = o3d.utility.Vector3dVector(rotated_vertices)
 
 
 class ICP:
