+++ src/kimera_vio_evaluation/tools/__init__.py
@@ -1,9 +1,7 @@
 """Tools and utilities for evaluation."""
 from kimera_vio_evaluation.tools.matplotlib_plotter import *
 from kimera_vio_evaluation.tools.plotly_plotter import *
-from kimera_vio_evaluation.tools.filesystem_utils import *
 from kimera_vio_evaluation.tools.latex_utils import *
 from kimera_vio_evaluation.tools.utils import *
-from kimera_vio_evaluation.tools.mesh import *
 from kimera_vio_evaluation.tools.website_utils import *
 from kimera_vio_evaluation.tools.trajectory_tools import *
